layout/generic/test/test_bug470212.html

<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=470212
-->
<head>
  <title>Test for Bug 470212 -  crash [@ nsContentUtils::ComparePoints]</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=470212">Mozilla Bug 470212</a>
<div style="width: 200px;">
<ca>
<canvas style="border: 1px solid black;"></canvas>
</ca>
</div>

<pre id="test">
<script>
function doShiftDrag(){
  setTimeout(function() {
    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
    var wu =  window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
                    .getInterface(Components.interfaces.nsIDOMWindowUtils);
    wu.sendMouseEvent('mousedown',  0, 50, 0, 1, 4);
    wu.sendMouseEvent('mousemove',  70, 70, 0, 0, 4);
    wu.sendMouseEvent('mousemove',  80, 500, 0, 0, 4);

    is(window.getSelection().rangeCount, 0, "rangeCount should be 0");
    wu.sendMouseEvent('mouseup',  80, 500, 0, 0, 4);

    SimpleTest.finish();
  }, 0);
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(doShiftDrag);
</script>

</pre>
</body>
</html>
